# Duration writer improvements