--- a/src/AvroConvert/AvroObjectServices/Write/Resolvers/Duration.cs
+++ b/src/AvroConvert/AvroObjectServices/Write/Resolvers/Duration.cs
@@ -16,6 +16,7 @@
 #endregion
 
 using System;
+using System.Runtime.InteropServices;
 using SolTechnology.Avro.AvroObjectServices.Schemas;
 using SolTechnology.Avro.Features.Serialize;
 using SolTechnology.Avro.Infrastructure.Exceptions;
@@ -30,24 +31,34 @@ namespace SolTechnology.Avro.AvroObjectServices.Write
             var duration = (TimeSpan)logicalValue;
 
             var baseSchema = (FixedSchema)schema.BaseTypeSchema;
-            byte[] bytes = new byte[baseSchema.Size];
-            var monthsBytes = BitConverter.GetBytes(0);
+            Span<byte> buffer = stackalloc byte[baseSchema.Size];
+            buffer.Slice(0, 4).Fill(0);
+
+#if NET6_0_OR_GREATER
+            var days = duration.Days;
+            var daysBytes = MemoryMarshal.AsBytes(MemoryMarshal.CreateSpan(ref days, 1));
+#else
             var daysBytes = BitConverter.GetBytes(duration.Days);
+#endif
+            daysBytes.CopyTo(buffer.Slice(4, 4));
 
             var milliseconds = ((duration.Hours * 60 + duration.Minutes) * 60 + duration.Seconds) * 1000 +
                                duration.Milliseconds;
+#if NET6_0_OR_GREATER
+            var millisecondsBytes = MemoryMarshal.AsBytes(MemoryMarshal.CreateSpan(ref milliseconds, 1));
+#else
             var millisecondsBytes = BitConverter.GetBytes(milliseconds);
-
-
-            System.Array.Copy(monthsBytes, 0, bytes, 0, 4);
-            System.Array.Copy(daysBytes, 0, bytes, 4, 4);
-            System.Array.Copy(millisecondsBytes, 0, bytes, 8, 4);
-
+#endif
+            millisecondsBytes.CopyTo(buffer.Slice(8, 4));
 
             if (!BitConverter.IsLittleEndian)
-                System.Array.Reverse(bytes); //reverse it so we get little endian.
+                buffer.Reverse();
+#if NET6_0_OR_GREATER
+            writer.WriteFixed(buffer);
+#else
+            writer.WriteFixed(buffer.ToArray());
+#endif
 
-            writer.WriteFixed(bytes);
         }
     }
 }