This PR reduces allocations and speeds up Duration write resolver.
Benchmark results of current vs optimized

| Method    | Mean     | Error    | StdDev   | Ratio | Gen0   | Allocated | Alloc Ratio |
|---------- |---------:|---------:|---------:|------:|-------:|----------:|------------:|
| Current   | 33.86 ns | 0.119 ns | 0.105 ns |  1.00 | 0.0325 |     136 B |        1.00 |
| Optimized | 12.45 ns | 0.241 ns | 0.226 ns |  0.37 |      - |         - |        0.00 |