@AdrianStrugala 2023-11-02
Amazing, thank you. I've started wondering if the `#use net 6 or greater` annotation is mandatory. Wouldn't setting up C# version on csproj level do the trick even for .NET Frameworks projects?