@gmanvel 2023-11-02
> Amazing, thank you. I've started wondering if the `#use net 6 or greater` annotation is mandatory. Wouldn't setting up C# version on csproj level do the trick even for .NET Frameworks projects?

@AdrianStrugala use of NET6 pre-processor is required because some of the Span based APIs are not supported for netstandaed 2.0 which `AvroConvert` supports.